add auto fetch after action within 5sec, add confirmation to actions
