+++ rubix-ui/frontend/src/components/hosts/views/constants.ts
@@ -1,0 +1,1 @@
+export const REFRESH_TIMEOUT = 5000; // 5 seconds
